Throw error if connections are made between different contexts
The WebAudio spec says that if a node is connected/disconnected to
another node or AudioParam that belongs to a different context, then
an InvalidStateError must be thrown.[1][2]  Add a check for this and throw
the required error.

This also requires a fix to audit.js which wasn't properly catching
DOM exceptions of the wrong type.

Finally, ctor-channelsplitter.html needed to be updated to specify the
error type because of the change in audit.js.

Also adds new WPT test, different-contexts.html, to more thoroughly test
connect/disconnects to different contexts.

[1] https://webaudio.github.io/web-audio-api/#dom-audionode-connect-destinationnode-output-input-destinationnode
[2] https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationnode-destinationnode

Bug: 1206927
Change-Id: I320425268d1fd243e347dd36c79172c69dcc9733