[LayoutNG] OOF with inline containing block in multicol
This CL adds support for fragmented OOFs with an inline containing
block. This involved determining the containing block info at the right
time and collecting all of the inline items from all fragmentainers in
order to compute the correct full containing block rect. Some of the
work involved in this update included:

1. Skipping over the call to ComputeInlineContainingBlocks() in
NGOutOfFlowLayoutPart::LayoutCandidates() when inside a fragmentation
context.

2. Adding a similar method to ComputeInlineContainingBlocks() called
ComputeInlineContainingBlocksForFragmentainer() that performs the
same calculation but in the case of fragmentation. This is called
in NGOutOfFlowLayoutPart::LayoutFragmentainerDescendants() in order
to set up the correct ContainingBlockInfo for the inline containing
block.

3. Move the existing inline containing block helpers from
NGBoxFragmentBuilder to a new utils file: InlineContainingBlockUtils.
One such helper, ComputeInlineContainerGeometry(), required different
logic in the case of fragmentation, so a new helper,
ComputeInlineContainerGeometryForFragmentainer(), was added to handle
the collection of inline items in this case. Besides the newly added
method, no changes were made to the methods moved from
NGBoxFragmentBuilder to InlineContainingBlockUtils.

4. Move PreviousFragmentainerBreakToken() from NGOutOfFlowLayoutPart
to NGBoxFragmentBuilder for easier access in InlineContainingBlockUtils.

A few follow-up changes include handling the relpos offset for inline
containing blocks, legacy writeback, and handling OOFs whose inline
containing block is a direct child of the multicol (CL:2938502).

Bug: 1079031
Change-Id: Ie5b46b7e49304d5c198c2df2e6a6735f9d2fc68c