Fix RecalcLayoutOverflow() for multicol in LayoutNGBlockFragmentation
The fragmentainer in multicol is still using a legacy layout object.
When we reach the fragment of the column, we get
a LayoutMultiColumnFlowThread legacy object,
and we call LayoutBlock::RecalcLayoutOverflow().
That means that we start to use the legacy layout code
to recalc the layout overflow.

To fix this this patch is skipping the fragmentainers
in RecalcChildLayoutOverflow(), and going to the children directly.
The fragmentainers have the positioned objects as children,
so we process them properly.

This only fixes the test when LayoutNGBlockFragmentation is enabled.

BUG=1221181
TEST=external/wpt/css/css-overflow/scrollable-overflow-transform-dynamic-007.html

Change-Id: I8ecda520551f2c3e832f2e77cdedcb1fc17fe99f