[LayoutNG] Multicol with OOF contained by inline ancestor
An OOF with an inline containing block that is a direct child of a
multicol crashed because the containing block of the inline element was
detected to be the LayoutFlowThread rather than the multicol. The reason
being that LayoutObject::FindNonAnonymousContainingBlock() checks
IsAnonymousBlock() when finding the containing block, which can be
a LayoutFlowThread. However, a LayoutFlowThread should not be used
as a containing block, so skip over these when finding the correct
containing block.

This results in different behavior in both legacy and NG than what
legacy produced previously. In the added wpt test, legacy would
break the abspos into two columns. However, with the new changes, the
multicol container is treated as the containing block, and the abspos
does not break. This matches Firefox's behavior, as well.

Bug: 1079031
Change-Id: I90dc78fc541abf2a93a62da808542787a3ad2e6f